A little refactor
