+++ src/main/java/com/evolvedbinary/jnibench/jmhbench/cache/ByteArrayCache.java
@@ -1,0 +1,76 @@
+/**
+ * Copyright © 2021, Evolved Binary Ltd
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *     * Redistributions of source code must retain the above copyright
+ *       notice, this list of conditions and the following disclaimer.
+ *     * Redistributions in binary form must reproduce the above copyright
+ *       notice, this list of conditions and the following disclaimer in the
+ *       documentation and/or other materials provided with the distribution.
+ *     * Neither the name of the <organization> nor the
+ *       names of its contributors may be used to endorse or promote products
+ *       derived from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL <COPYRIGHT HOLDER> BE LIABLE FOR ANY
+ * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package com.evolvedbinary.jnibench.jmhbench.cache;
+
+import java.nio.ByteBuffer;
+import java.nio.LongBuffer;
+
+public class ByteArrayCache extends LinkedListAllocationCache<byte[]> {
+
+    @Override
+    byte[] allocate(int valueSize) {
+        return new byte[valueSize];
+    }
+
+    @Override
+    void free(byte[] buffer) {
+        //nothing to do
+    }
+
+
+    @Override
+    public byte[] copyOut(byte[] item) {
+
+        byte[] array = byteArrayOfSize(item.length);
+        ByteBuffer target = ByteBuffer.wrap(array);
+        target.put(item);
+
+        return array;
+    }
+
+    @Override
+    public int byteChecksum(byte[] item) {
+
+        int sum = 0;
+        for (int i = 0; i < item.length; i++) {
+            sum += item[i];
+        }
+
+        return sum;
+    }
+
+    @Override
+    public int longChecksum(byte[] item) {
+        LongBuffer buffer = ByteBuffer.wrap(item).asLongBuffer();
+        long sum = 0;
+        while (buffer.remaining() > 0) {
+            sum += buffer.get();
+        }
+        return (int)sum;
+    }
+}
+
