+++ src/main/java/com/evolvedbinary/jnibench/jmhbench/cache/LinkedListAllocationCache.java
@@ -24,8 +24,9 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.evolvedbinary.jnibench.jmhbench.common;
+package com.evolvedbinary.jnibench.jmhbench.cache;
 
+import com.evolvedbinary.jnibench.jmhbench.cache.AllocationCache;
 import org.openjdk.jmh.infra.Blackhole;
 
 import java.util.HashMap;
